feat: add reachable tile search to pathfinder
